+++ packages/mock/src/handlers/TestError.ts
@@ -1,17 +1,11 @@
-import { ChatInputCommandInteraction, MessageFlags } from 'discord.js';
+import { ChatInputCommandInteraction } from 'discord.js';
 import { Catchable, InteractionHandler, SlashRoute } from 'seedcord';
 
 @SlashRoute('throw')
 export class TestError extends InteractionHandler<ChatInputCommandInteraction> {
   @Catchable()
   async execute(): Promise<void> {
-    // wait for 5 seconds to throw Unknown Interaction Error
-    await new Promise((resolve) => setTimeout(resolve, 5000));
-    await this.event.reply({
-      content: 'This is a test error',
-      flags: MessageFlags.Ephemeral
-    });
-
-    this.core.db.services.users.test();
+    await Promise.resolve();
+    throw new Error('Test error');
   }
 }