+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @seedcord/types
 
+## 0.1.4
+
+### Patch Changes
+
+- 8a7591a: bump deps
+
 ## 0.1.3
 
 ### Patch Changes